implemented user registration
