+++ AccountsManager/templates/AccountsManager/base.html
@@ -1,0 +1,17 @@
+{% extends "main/base.html" %}
+{% load crispy_forms_tags %}
+
+{% block content %}
+    {% block heading %}
+    {% endblock heading %}
+    <form class="form" method="post" action="/register/">
+        <div>
+            {% block form %}
+                {% csrf_token %}
+                {{ form|crispy }}
+            {% endblock form %}
+            {% block button %}
+            {% endblock button %}
+        </div>
+    </form>
+{% endblock content %}
