+++ AccountsManager/templates/AccountsManager/register.html
@@ -1,14 +1,15 @@
-{% extends "main/base.html" %}
+{% extends "AccountsManager/base.html" %}
+
 {% block title %}
 Weather: Register
 {% endblock title %}
     <body>
-        <form method="post" action="/login/">
+        <form method="post" action="/register/">
             {% block heading %}
-            <h1 class="mt-2 mb-4"> Login </h1>
+            <h1 class="mt-2 mb-4"> Register </h1>
             {% endblock heading %}
             {% block button %}
-            <button type="submit" name="login" class="btn btn-primary">Login</button>
+            <button type="submit" name="register" class="btn btn-primary">Register</button>
             {% endblock button %}
         </form>
     </body>
