+++ AccountsManager/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("", views.home, name="home"),
     path("register/", views.register, name="register"),
     path("login/", views.login, name="login"),
 ]
