+++ main/templates/main/base.html
@@ -45,19 +45,7 @@
         <div class="main " id="content" name="content">
             <div class="row justify-content-center">
                 <div class="col-4">
-                    {% block heading %}
-                    {% endblock heading %}
                     {% block content %}
-                        <form class="form" method="post" action="/register/">
-                            <div>
-                                {% block form %}
-                                    {% csrf_token %}
-                                    {{ form.as_p }}
-                                {% endblock form %}
-                                {% block button %}
-                                {% endblock button %}
-                            </div>
-                        </form>
                     {% endblock content %}
                 </div>
             </div>
